refactor(layout): move all header logic into an isolated component
